WIP: FlaggingViewController
+ Refactored and added marks
+ UITableViewDelegate protocol function implementation
